Update signing-up.md
I made a change to a typographical error on the "Signing Up" doc page, under the "Verifying your number" Heading. The typographical error can be found in this statement: 'If you see the "We need you to verify your phone number to complete your sign-up" banner, click the "Verify your numer now" link to access the service.' Letter 'b' in "Verify your number" was omitted.