kubernetes-volumes - re-work example output
Spinning up a clean cluster, the `civo-volume` storage class has
`WaitForFirstConsumer` rather than `Immediate` volume binding.

This means the volume is not created when the volume claim is created,
but rather when the pod is scheduled.

The `get pv` output is moved to after the `get pods` output, which will
successfully show the relevant volume.